New Functionality: Build New DuckDB File
This feature takes an existing DuckDB file (as referenced in the config.py file), and performs update & append job types for correctly formatted .csv files in the latest timestamped *_update folder.

Errors are recorded in an errors.json file. Large csv files are processed in chunks of 1,000 rows, and data must match the schema of the existing table. The feature also includes date format normalization, which resolves some issues with Excel automatically changing date formats if you open the csv in Excel.

Update jobs reference the fields denoted for each table in the filtering_criteria.json file. Job types are denoted by the csv file name pattern. If a row in an update job fails to find an existing match based on the defined filtering criteria, the row will be appended instead. Note: in an update job, matched rows are rewritten based on the provided data for the row in the update job.